requirements.txt
graphviz
GitPython